intention dialog flow coming together
