+++ src/applescript/dialog.js
@@ -21,16 +21,17 @@ function renderButtons(buttons) {
 }
 
 async function showDialog(title, text, buttons, givingUpAfter = 30) {
-    const btns = renderButtons(buttons)
+
+    const defaultButton = buttons.slice(-1)[0]  // grab the right most button
+    buttons = buttons.map(b=> `"${b}"`).join(", ")
 
     const SCRIPT = ` 
 
-    ${btns.defs}
     set result to ¬ 
         (display dialog "${text}" ¬ 
         with title "${title}" ¬
-        buttons { ${btns.refs} } ¬
-        default button button1Label ¬ 
+        buttons { ${buttons} } ¬
+        default button "${defaultButton}"¬ 
         giving up after ${givingUpAfter} ¬
         with icon alias ¬
             ((path to application support from user domain as text) & "com.focusbear.latenomore:icon.png"))
@@ -41,17 +42,16 @@ async function showDialog(title, text, buttons, givingUpAfter = 30) {
 }
 
 async function askQuestion(question, title, buttons, defaultButton) {
-    const btns = renderButtons(buttons)
 
-    if (!defaultButton) defaultButton = buttons[0]
-
+    if (!defaultButton) defaultButton = buttons.slice(-1)[0]  // grab the right most button
     buttons = buttons.map(b=> `"${b}"`).join(", ")
+
     const SCRIPT = `
 
-  --  try    
+    try    
         set result to ¬
             (display dialog "${question}"¬
-            default answer "" ¬
+            default answer "\n\n\n\n" ¬
             with title "${title}" ¬
             buttons { ${buttons} }  ¬
             default button "${defaultButton}")
@@ -59,7 +59,7 @@ async function askQuestion(question, title, buttons, defaultButton) {
         set buttonReturned to button returned of result
         set userInput to text returned of hold
         return {buttonReturned, userInput}
- --   end try`
+   end try`
 
     const [buttonReturned, userInput] = await exec(SCRIPT)
     return { buttonReturned, userInput }
