+++ src/events/index.js
@@ -22,6 +22,9 @@ function remove(listName, evt) {
         ({ id, type }) => evt.id !== id && evt.type !== type
     )
 }
+function has(listName, evt){
+    return listName.filter(({id, type})=> id == evt.id && type == evt.type)
+}
 
 function get(listName) {
     if (!listName) return events
@@ -33,4 +36,5 @@ module.exports = {
     add,
     remove,
     get,
+    has
 }
